assignment-04-DA24M011-Nandhakishore/news_fetcher/requirements.txt
requests
psycopg2
feedparser
pillow
